requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
apprise==1.3.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
Markdown==3.4.3
multidict==6.0.4
oauthlib==3.2.2
pmaw==3.0.0
praw==7.7.0
prawcore==2.3.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
sniffio==1.3.0
soupsieve==2.4.1
tqdm==4.65.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
yarl==1.9.2
